Add CI/CD pipeline and in-memory DB configuration
Added a GitHub Actions workflow to automate build and deployment
of the `public-api` application to Azure Web App. The workflow
includes build and deploy jobs, with environment variables for
reuse.

Updated `appsettings.json` to include the
`"UseOnlyInMemoryDatabase": true` configuration for enabling an
in-memory database.

Modified `ServiceCollectionExtensions.cs` to always configure
the local database context by altering the conditional logic.